+++ src/main/kotlin/br/com/validation/request/validationRequest/controller/response/QrCodeResponse.kt
@@ -1,11 +1,10 @@
 package br.com.validation.request.validationRequest.controller.response
 
-import java.math.BigDecimal
 import java.time.LocalDateTime
 
 data class QrCodeResponse (
 
     val transactionId: String,
-    val amount: BigDecimal,
+    val amount: Double,
     val createdAt: LocalDateTime
 )
